backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
sse-starlette==1.8.2
